react/leaflet: add references and memoization
Geojson wrapper is rerendered each time state properties are modified.
Leaflet React component will reset the style of the entire map  when
we hover on a geojson. So when we call callback that changes the
style of the geojson, it will also trigger the rerendering of the
Geojson react component afterwards and resets all styles.

To prevent rerendering of the leaflet Geojson component, we add callback
and ref memoization of component props.