choropleth: Fix map styling when changing scale
After modifying scale, geojson  map is not rerendered.

To fix this, use a memo with a function settting update when scahle
changes. We need this because, we will rerender unecessarely when modifying
props on parent component. We do not want this because styling will reset
onMouseOver callback function.

We need also a key to force unmouting and init a new leaflet component
because when changing scale, resetting style during ouMouseOut cb will
reset to the default scale and not the new one.