choropleth: Fix map legend information reactivity
This is a follow-up of 9669a7b6eefd84c25ecddfa996c27ae4efd5546b
Legend information reactivity was not fixed correctly by the above commit.

Before, the mouseOver event does no trigger a rerendering of the legend
information scale.
So in this commit, we prevent the Geojson
rerendering but not the choropleth rendering.